requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze = requirements.txt' in the terminal to overwrite this file
Django>=3.2.15
gunicorn>=20.1.0
django-multiselectfield>=0.1.12
django-crispy-forms>=1.14.0
crispy-bootstrap5>=0.7
envconfiguration>=0.1.0